refactoring the UserController, adding the user view
