+++ src/Form/PasswordUpdateType.php
@@ -7,7 +7,6 @@ use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\EqualTo;
 use Symfony\Component\Validator\Constraints\Length;
 
 class PasswordUpdateType extends AbstractType
@@ -22,6 +21,7 @@ class PasswordUpdateType extends AbstractType
             ->add('password', RepeatedType::class, [
                 'type' => PasswordType::class,
                 'invalid_message' => "Vous n'avez pas correctement confirmé votre mot de passe.",
+                'constraints' => new Length(['min' => 6, 'minMessage' => 'Le mot de passe doit contenir six caractères minimum.']),
                 'required' => true,
                 'first_options' => ['label' => 'Nouveau mot de passe', 'attr' => ['placeholder' => 'Taper votre nouveau mot de passe...']],
                 'second_options' => ['label' => 'Confirmation mot de passe', 'attr' => ['placeholder' => 'Confirmer votre nouveau mot de passe..']]
