+++ symfony.lock
@@ -302,6 +302,15 @@
     "symfony/yaml": {
         "version": "v4.2.1"
     },
+    "twig/extensions": {
+        "version": "1.0",
+        "recipe": {
+            "repo": "github.com/symfony/recipes",
+            "branch": "master",
+            "version": "1.0",
+            "ref": "ddb2e0a77773b7fd75d8d649545f174e664500ab"
+        }
+    },
     "twig/twig": {
         "version": "v2.6.0"
     },
