+++ templates/base.html.twig
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>Snowtricks - {% block title %}Le site{% endblock %}</title>
     <link rel="stylesheet" href="/css/bootstrap.min.css">
+    <link rel="stylesheet" href="/css/app.css">
     {% block stylesheets %}{% endblock %}
 </head>
 <body>
