+++ templates/partials/header.html.twig
@@ -13,14 +13,14 @@
         <ul class="navbar-nav ml-auto">
             {% if app.user %}
                 <li class="nav-item">
-                    <a href="{{ path('user_logout') }}" class="nav-link">Deconnexion</a>
+                    <a href="{{ path('security_logout') }}" class="nav-link">Deconnexion</a>
                 </li>
             {% else %}
                 <li class="nav-item">
-                    <a href="{{ path('user_register') }}" class="nav-link">Inscription</a>
+                    <a href="{{ path('security_register') }}" class="nav-link">Inscription</a>
                 </li>
                 <li class="nav-item">
-                    <a href="{{ path('user_login') }}" class="nav-link">Connexion</a>
+                    <a href="{{ path('security_login') }}" class="nav-link">Connexion</a>
                 </li>
 
             {% endif %}
