--- templates/user/login.html.twig
@@ -1,20 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Page de connexion{% endblock %}
-
-{% block body %}
-    <div class="container">
-        <div class="row">
-            <div class="col-md-3">&nbsp;</div>
-            <div class="col-md-6">
-                <div class="bg-light py-3 px-3">
-                    <h1>Connectez vous !</h1>
-                    {{ form_start(form) }}
-                    {{ form_widget(form) }}
-                    <button type="submit" class="btn btn-primary">Connexion</button>
-                    {{ form_end(form) }}
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
