+++ templates/user/show.html.twig
@@ -1,0 +1,34 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}
+    Profil
+    {% if elision(user.fullname) %}
+        d'{{ user.fullname }}
+    {% else %}
+        de {{ user.fullname }}
+    {% endif %}
+{% endblock %}
+
+{% block body %}
+    <div class="container">
+        <div class="row align-items-center">
+            <div class="col-3 text-center">
+                <img src="{{ user.picture }}" class="avatar mb-3" alt="Avatar de {{ user.fullName }}">
+            </div>
+            <div class="col-9">
+                <h1>{{ user.fullname|capitalize }}</h1>
+                {% if user is same as (app.user) %}
+                    <div class="mt-3">
+                        <a href="{{ path('user_profile') }}" class="btn btn-primary mr-2 mb-2">Modifier mes informations</a>
+                        <a href="{{ path('security_password_update') }}" class="btn btn-primary mr-2 mb-2">Modifier mon mot de passe</a>
+                    </div>
+                {% endif %}
+                {{ user.introduction }}
+            </div>
+            <hr width="75%">
+            <div class="col-12">
+                {{ user.description }}
+            </div>
+        </div>
+    </div>
+{% endblock %}
